+++ app/src/androidTest/java/it/units/borghisegreti/utils/CustomMatchers.java
@@ -1,4 +1,4 @@
-package it.units.borghisegreti;
+package it.units.borghisegreti.utils;
 
 import android.view.View;
