+++ firebase-emulators-data/auth_export/accounts.json
@@ -1,1 +1,29 @@
-{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"RCpeuPJSbvWjWpdVz3mw1q2j09Cz","lastLoginAt":"1688914378387","displayName":"","photoUrl":"","emailVerified":false,"email":"test@prova.it","salt":"fakeSaltLJ3lckqvA8e9LZlC1HEp","passwordHash":"fakeHash:salt=fakeSaltLJ3lckqvA8e9LZlC1HEp:password=password","passwordUpdatedAt":1688914378388,"validSince":"1688914378","mfaInfo":[],"createdAt":"1688914378387","providerUserInfo":[{"providerId":"password","email":"test@prova.it","federatedId":"test@prova.it","rawId":"test@prova.it","displayName":"","photoUrl":""}]}]}
+{
+  "kind": "identitytoolkit#DownloadAccountResponse",
+  "users": [
+    {
+      "localId": "RCpeuPJSbvWjWpdVz3mw1q2j09Cz",
+      "lastLoginAt": "1688914378387",
+      "displayName": "",
+      "photoUrl": "",
+      "emailVerified": false,
+      "email": "test@prova.it",
+      "salt": "fakeSaltLJ3lckqvA8e9LZlC1HEp",
+      "passwordHash": "fakeHash:salt=fakeSaltLJ3lckqvA8e9LZlC1HEp:password=password",
+      "passwordUpdatedAt": 1688914378388,
+      "validSince": "1688914378",
+      "mfaInfo": [],
+      "createdAt": "1688914378387",
+      "providerUserInfo": [
+        {
+          "providerId": "password",
+          "email": "test@prova.it",
+          "federatedId": "test@prova.it",
+          "rawId": "test@prova.it",
+          "displayName": "",
+          "photoUrl": ""
+        }
+      ]
+    }
+  ]
+}
